backend/requirements.txt
# Core Framework
fastapi>=0.115.2
uvicorn[standard]==0.27.0
pydantic>=2.7.4
pydantic-settings>=2.4.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# LightRAG
lightrag-hku>=1.4.9
nano-vectordb>=0.0.4

# LLM & Embeddings
openai>=1.10.0
google-generativeai>=0.3.2
ollama>=0.1.6
sentence-transformers>=2.3.1

# Document Processing
pypdf>=4.0.1
python-docx>=1.1.0
python-multipart>=0.0.6

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
loguru>=0.7.2

# Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
black>=24.1.1
ruff>=0.1.14
